backend/pyproject.toml
[project]
name = "hyperion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
authors = [
    { name = "Arian Ott", email = "arian.ott@ieee.org" }
]
license = { text = "GPL-3.0-or-later" }
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
]
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncmy>=0.2.10",
    "faker>=39.0.0",
    "fastapi[all]>=0.126.0",
    "orjson>=3.11.5",
    "pwdlib>=0.3.0",
    "pyartnet>=2.0",
    "pydantic-settings>=2.12.0",
    "pygount>=3.1.0",
    "pyjwt>=2.10.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]


[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.10",
]
